Correct input param doc, update output params
Corrected the description of the input parameter "pathPlanes" from "an array of Points" to "an array of Planes". Commented out the "SuperShape" output parameter and added a new "SMT_SuperShapes" parameter using `GH_SuperShapeParameter` with `GH_ParamAccess.tree`. Also commented out the line setting data for the "SuperShape" output parameter.